support NO_REVOKE

+++ letest.sh
@@ -893,9 +893,10 @@ le_test_standandalone_renew_v2() {
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
 
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
-
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   rm -rf "$certdir"
   mkdir -p "$certdir"
   
@@ -910,9 +911,10 @@ le_test_standandalone_renew_v2() {
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
   
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
-
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   rm -rf "$certdir"
   
 
@@ -945,8 +947,11 @@ le_test_standandalone_renew_localaddress_v2() {
   _assertfileequals "$lehome/$TestingDomain/$TestingDomain.key" "$key" ||  return
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   rm -rf "$certdir"
   mkdir -p "$certdir"
   
@@ -965,8 +970,11 @@ le_test_standandalone_renew_localaddress_v2() {
   _assertfileequals "$lehome/$TestingDomain/$TestingDomain.key" "$key" ||  return
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   rm -rf "$certdir"
   
 
@@ -999,8 +1007,11 @@ le_test_standandalone_listen_v4_v2() {
   _assertfileequals "$lehome/$TestingDomain/$TestingDomain.key" "$key" ||  return
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   rm -rf "$certdir"
   mkdir -p "$certdir"
   
@@ -1020,8 +1031,10 @@ le_test_standandalone_listen_v4_v2() {
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
   
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   rm -rf "$certdir"
   
 
@@ -1060,8 +1073,11 @@ le_test_standandalone_listen_v6_v2() {
   _assertfileequals "$lehome/$TestingDomain/$TestingDomain.key" "$key" ||  return
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   rm -rf "$certdir"
   mkdir -p "$certdir"
   
@@ -1080,8 +1096,12 @@ le_test_standandalone_listen_v6_v2() {
   _assertfileequals "$lehome/$TestingDomain/$TestingDomain.key" "$key" ||  return
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
+  
   rm -rf "$certdir"
   
 
@@ -1114,8 +1134,11 @@ le_test_standandalone_deactivate_v2() {
   _assertfileequals "$lehome/$TestingDomain/$TestingDomain.key" "$key" ||  return
   _assertfileequals "$lehome/$TestingDomain/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   rm -rf "$certdir"
   mkdir -p "$certdir"
   
@@ -1146,9 +1169,11 @@ le_test_standandalone_SAN() {
   _assertcmd "$lehome/$PROJECT_ENTRY  --server \"$TEST_ACME_Server\"  --issue -d \"$TestingDomain\" -d \"$TestingAltDomains\" --standalone" ||  return
   _assertcert "$lehome/$TestingDomain/$TestingDomain.cer" "$TestingDomain" "$CA" || return
   _assertcert "$lehome/$TestingDomain/ca.cer" "$CA" || return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
-
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
   
   
 }
@@ -1180,8 +1205,11 @@ le_test_standandalone_ECDSA_256_renew() {
   _assertcmd "$lehome/$PROJECT_ENTRY  --server \"$TEST_ACME_Server\"  --issue -d $TestingDomain --standalone -k ec-256" ||  return
   _assertcert "$lehome/$TestingDomain$ECC_SUFFIX/$TestingDomain.cer" "$TestingDomain" "$CA_ECDSA" || return
   _assertcert "$lehome/$TestingDomain$ECC_SUFFIX/ca.cer" "$CA_ECDSA" || return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain --ecc" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain --ecc" ||  return
+  fi
 
   sleep 5
   _assertcmd "$lehome/$PROJECT_ENTRY --renew -d $TestingDomain --force --ecc" ||  return
@@ -1224,8 +1252,10 @@ le_test_standandalone_ECDSA_256_SAN_renew_v2() {
   _assertfileequals "$lehome/$TestingDomain$ECC_SUFFIX/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain$ECC_SUFFIX/fullchain.cer" "$full" ||  return
   
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain --ecc" ||  return
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain --ecc" ||  return
+  fi
   
   rm -rf "$certdir"
   mkdir -p "$certdir"
@@ -1239,9 +1269,11 @@ le_test_standandalone_ECDSA_256_SAN_renew_v2() {
   _assertfileequals "$lehome/$TestingDomain$ECC_SUFFIX/$TestingDomain.key" "$key" ||  return
   _assertfileequals "$lehome/$TestingDomain$ECC_SUFFIX/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$TestingDomain$ECC_SUFFIX/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain --ecc"||  return
   
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain --ecc"||  return
+  fi
 
 
 }
@@ -1279,9 +1311,11 @@ le_test_standandalone_ECDSA_384() {
   _assertcmd "$lehome/$PROJECT_ENTRY --server \"$TEST_ACME_Server\"  --issue --standalone -d \"$TestingDomain\" -k ec-384" ||  return
   _assertcert "$lehome/$TestingDomain$ECC_SUFFIX/$TestingDomain.cer" "$TestingDomain" "$CA_ECDSA" || return
   _assertcert "$lehome/$TestingDomain$ECC_SUFFIX/ca.cer" "$CA_ECDSA" || return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain --ecc" ||  return
-
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain --ecc" ||  return
+  fi
 
 
 }
@@ -1318,8 +1352,12 @@ le_test_standandalone_renew_idn_v2() {
   _assertfileequals "$lehome/$_test_idn/$_test_idn.key" "$key" ||  return
   _assertfileequals "$lehome/$_test_idn/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$_test_idn/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $_test_idn" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $_test_idn" ||  return
+  fi
+  
   rm -rf "$certdir"
   mkdir -p "$certdir"
   
@@ -1333,8 +1371,12 @@ le_test_standandalone_renew_idn_v2() {
   _assertfileequals "$lehome/$_test_idn/$_test_idn.key" "$key" ||  return
   _assertfileequals "$lehome/$_test_idn/ca.cer" "$ca" ||  return
   _assertfileequals "$lehome/$_test_idn/fullchain.cer" "$full" ||  return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $_test_idn" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $_test_idn" ||  return
+  fi
+  
   rm -rf "$certdir"
   
 
@@ -1370,8 +1412,11 @@ le_test_dnsapi() {
     _assertcert "$lehome/$TestingDomain/$TestingDomain.cer" "$TestingDomain" "$CA" || return
     _assertcert "$lehome/$TestingDomain/ca.cer" "$CA" || return
     _assertcmd "$lehome/$PROJECT_ENTRY --deactivate -d \"$TestingDomain\" >/dev/null 2>&1"
-    sleep 5
-    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+    
+    if [ -z "$NO_REVOKE" ]; then
+      sleep 5
+      _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+    fi
     ) || return
   else
     (
@@ -1389,8 +1434,11 @@ le_test_dnsapi() {
     _assertcert "$lehome/$TestingDomain/$TestingDomain.cer" "$TestingDomain" "$CA" || return
     _assertcert "$lehome/$TestingDomain/ca.cer" "$CA" || return
     _assertcmd "$lehome/$PROJECT_ENTRY --deactivate -d \"$TestingDomain\" >/dev/null 2>&1"
-    sleep 5
-    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+    
+    if [ -z "$NO_REVOKE" ]; then
+      sleep 5
+      _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+    fi
     ) || return 
   fi
 
@@ -1422,8 +1470,11 @@ le_test_standandalone_ipcert() {
   _assertcmd "$lehome/$PROJECT_ENTRY --server \"$TEST_ACME_Server\"  --issue -d $TestingDomain --standalone" ||  return
   _assertcert "$lehome/$TestingDomain/$TestingDomain.cer" "$TestingDomain" "$CA" || return
   _assertcert "$lehome/$TestingDomain/ca.cer" "$CA" || return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
 
 
 }
@@ -1454,8 +1505,10 @@ le_test_alpn_ipcert() {
   _assertcert "$lehome/$TestingDomain/$TestingDomain.cer" "$TestingDomain" "$CA" || return
   _assertcert "$lehome/$TestingDomain/ca.cer" "$CA" || return
 
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
 
 }
 
@@ -1480,8 +1533,11 @@ le_test_preferred_chain() {
   _assertcmd "$lehome/$PROJECT_ENTRY  --server \"$TEST_ACME_Server\"  --issue -d \"$TestingDomain\" -d \"$TestingAltDomains\" --standalone --preferred-chain \"$TEST_PREFERRED_CHAIN\" " ||  return
   _assertcert "$lehome/$TestingDomain/$TestingDomain.cer" "$TestingDomain" "$CA" || return
   _assertcert "$lehome/$TestingDomain/ca.cer" "$CA" "$TEST_PREFERRED_CHAIN" || return
-  sleep 5
-  _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  
+  if [ -z "$NO_REVOKE" ]; then
+    sleep 5
+    _assertcmd "$lehome/$PROJECT_ENTRY --revoke -d $TestingDomain" ||  return
+  fi
 
 
 }
